Remove all references to legacy sources and source actors
