+++ src/devtools/client/debugger/src/actions/sources/newSources.ts
@@ -12,57 +12,26 @@
 import type { UIThunkAction } from "ui/actions";
 import { fetchPossibleBreakpointsForSource } from "ui/reducers/possibleBreakpoints";
 
-import { insertSourceActors } from "../../actions/source-actors";
-import { makeSourceId } from "../../client/create";
 import { setRequestedBreakpoint } from "../../reducers/breakpoints";
 import type { Context } from "../../reducers/pause";
-import { SourceActor } from "../../reducers/source-actors";
-import {
-  getBlackBoxList,
-  hasSourceActor,
-  getPendingBreakpointsForSource,
-  getContext,
-} from "../../selectors";
+import { getPendingBreakpointsForSource, getContext } from "../../selectors";
 import { getTabs, tabsRestored } from "devtools/client/debugger/src/reducers/tabs";
-import { getPendingBreakpointList } from "devtools/client/debugger/src/reducers/pending-breakpoints";
 
 import {
   getSourceDetails,
   allSourcesReceived,
   SourceDetails,
   getAllSourceDetails,
-  getSourceDetailsEntities,
-  getSourceByUrl,
   getSourceToDisplayForUrl,
   getSourceToDisplayById,
 } from "ui/reducers/sources";
-import { getRawSourceURL } from "../../utils/source";
 import { syncBreakpoint } from "../breakpoints";
 
 import { toggleBlackBox } from "./blackbox";
 import { experimentalLoadSourceText, getPreviousPersistedLocation } from "ui/reducers/sources";
 import { AppStartListening } from "ui/setup/listenerMiddleware";
 
-interface SourceData {
-  source: {
-    actor: string;
-    sourceMapURL?: string;
-    url: string;
-    extensionName?: string;
-    introductionUrl?: string;
-    introductionType?: string;
-    isBlackBoxed?: boolean;
-  };
-  thread: string;
-  id?: string;
-  url?: string;
-  isServiceWorker?: boolean;
-}
-
-interface SourceInfo {
-  type: "generated" | "original";
-  data: SourceData;
-}
+import { prefs } from "../../utils/prefs";
 
 // If a request has been made to show this source, go ahead and
 // select it.
@@ -131,6 +100,11 @@ function checkPendingBreakpoints(cx: Context, sourceId: string): UIThunkAction {
   };
 }
 
+export function getBlackBoxList(): string[] {
+  // @ts-expect-error Check on actual prefs usage here
+  return prefs.tabsBlackBoxed || [];
+}
+
 function restoreBlackBoxedSources(cx: Context, sources: SourceDetails[]): UIThunkAction {
   return async dispatch => {
     const tabs = getBlackBoxList();
@@ -145,186 +119,6 @@ function restoreBlackBoxedSources(cx: Context, sources: SourceDetails[]): UIThun
       }
     }
     */
-  };
-}
-
-// TODO Delete this!
-export function newQueuedSources(sourceInfo: SourceInfo[]): UIThunkAction<Promise<void>> {
-  return async dispatch => {
-    const generated: SourceData[] = [];
-    const original: SourceData[] = [];
-    for (const source of sourceInfo) {
-      if (source.type === "generated") {
-        generated.push(source.data);
-      } else {
-        original.push(source.data);
-      }
-    }
-
-    if (generated.length > 0) {
-      await dispatch(newGeneratedSources(generated));
-    }
-    if (original.length > 0) {
-      await dispatch(newOriginalSources(original));
-    }
-  };
-}
-
-// TODO Delete this!
-export function newOriginalSource(sourceInfo: SourceData): UIThunkAction<Promise<SourceDetails>> {
-  return async dispatch => {
-    const sources = await dispatch(newOriginalSources([sourceInfo]));
-    return sources[0];
-  };
-}
-
-// TODO Delete this!
-export function newOriginalSources(
-  sourceInfo: SourceData[]
-): UIThunkAction<Promise<SourceDetails[]>> {
-  return async (dispatch, getState) => {
-    const state = getState();
-    const seen = new Set();
-    const sources: SourceDetails[] = [];
-
-    // TODO Rewrite "check pending" handling
-    /*
-    for (const { id, url } of sourceInfo) {
-      if (seen.has(id) || getSourceDetails(state, id!)) {
-        continue;
-      }
-
-      seen.add(id);
-
-      sources.push({
-        extensionName: null,
-        id: id!,
-        introductionType: undefined,
-        introductionUrl: null,
-        isBlackBoxed: false,
-        isExtension: false,
-        isOriginal: true,
-        isPrettyPrinted: false,
-        relativeUrl: url,
-        url,
-      });
-    }
-
-    const cx = getContext(state);
-    dispatch(addSources(cx, sources));
-
-    await dispatch(checkNewSources(cx, sources));
-
-    for (const source of sources) {
-      dispatch(checkPendingBreakpoints(cx, source.id));
-    }
-*/
-    return sources;
-  };
-}
-
-export function newGeneratedSource(sourceInfo: SourceData): UIThunkAction<Promise<SourceDetails>> {
-  return async dispatch => {
-    const sources = await dispatch(newGeneratedSources([sourceInfo]));
-    return sources[0];
-  };
-}
-export function newGeneratedSources(
-  sourceInfo: SourceData[]
-): UIThunkAction<Promise<SourceDetails[]>> {
-  return async (dispatch, getState, { ThreadFront }) => {
-    const resultIds: string[] = [];
-
-    // TODO Rewrite "check pending" handling
-    /*
-    const newSourcesObj: Record<string, Source> = {};
-    const newSourceActors: SourceActor[] = [];
-
-    for (const { thread, isServiceWorker, source, id } of sourceInfo) {
-      const newId: string = id || makeSourceId(source, isServiceWorker);
-
-      const kind = ThreadFront.getSourceKind(source.actor);
-      const isPrettyPrinted = kind == "prettyPrinted";
-      const isOriginal = kind == "sourceMapped";
-
-      let url: string | undefined = source.url;
-      if (kind == "inlineScript") {
-        // Ignore inline scripts. We should see an HTML page script that includes
-        // these plus the rest of the HTML.
-        url = undefined;
-      }
-
-      if (ThreadFront.isMinifiedSource(source.actor)) {
-        // Ignore minified sources which have a pretty printed version.
-        url = undefined;
-      }
-
-      if (!getSourceDetails(getState(), newId) && !newSourcesObj[newId]) {
-        newSourcesObj[newId] = {
-          extensionName: source.extensionName,
-          id: newId,
-          introductionType: source.introductionType,
-          introductionUrl: source.introductionUrl,
-          isBlackBoxed: false,
-          isExtension: false,
-          isOriginal,
-          isPrettyPrinted,
-          relativeUrl: url,
-          url,
-        };
-      }
-
-      // We are sometimes notified about a new source multiple times if we
-      // request a new source list and also get a source event from the server.
-      if (!hasSourceActor(getState(), source.actor)) {
-        newSourceActors.push({
-          actor: source.actor,
-          id: source.actor,
-          introductionType: source.introductionType,
-          introductionUrl: source.introductionUrl,
-          isBlackBoxed: source.isBlackBoxed,
-          source: newId,
-          sourceMapURL: source.sourceMapURL,
-          thread,
-          url: source.url,
-        });
-      }
-
-      resultIds.push(newId);
-    }
-
-    const newSources = Object.values(newSourcesObj);
-
-    const cx = getContext(getState());
-    dispatch(addSources(cx, newSources));
-    dispatch(insertSourceActors(newSourceActors));
-
-    await dispatch(checkNewSources(cx, newSources));
-
-    for (const { source } of newSourceActors) {
-      dispatch(checkPendingBreakpoints(cx, source));
-    }
-    */
-    return resultIds.map(id => getSourceDetails(getState(), id)!);
-  };
-}
-
-/*
-function addSources(cx: Context, sources: Source[]): UIThunkAction {
-  return dispatch => {
-    dispatch({ cx, sources, type: "ADD_SOURCES" });
-  };
-}
-*/
-
-function checkNewSources(cx: Context, sources: SourceDetails[]): UIThunkAction {
-  return async (dispatch, getState) => {
-    for (const source of sources) {
-      dispatch(checkSelectedSource(cx, source));
-    }
-
-    dispatch(restoreBlackBoxedSources(cx, sources));
-    return sources;
   };
 }
