+++ src/devtools/client/debugger/src/actions/ui.ts
@@ -5,7 +5,6 @@
 //
 
 import type { Context } from "devtools/client/debugger/src/reducers/pause";
-// @ts-ignore no definition
 import { getCodeMirror } from "devtools/client/debugger/src/utils/editor";
 import type { UIThunkAction } from "ui/actions";
 import { getSelectedPanel, getToolboxLayout } from "ui/reducers/layout";
@@ -35,7 +34,6 @@ import {
 } from "ui/reducers/sources";
 import { isFulfilled } from "../utils/async-value";
 import { copyToTheClipboard } from "../utils/clipboard";
-// @ts-ignore no definition
 import { getEditor, getLocationsInViewport } from "../utils/editor";
 import { resizeBreakpointGutter } from "../utils/ui";
