--- src/devtools/client/debugger/src/components/App.d.ts
@@ -1,1 +1,0 @@
-export default function DebuggerLoader(): JSX.Element;
