--- src/devtools/client/debugger/src/reducers/index.d.ts
@@ -1,7 +1,0 @@
-import { Reducer } from "redux";
-
-declare const _default: {
-  [slice: string]: Reducer;
-};
-
-export default _default;
