+++ src/devtools/client/debugger/src/reducers/index.ts
@@ -9,8 +9,6 @@
  * @module reducers/index
  */
 
-import sourceActors from "./source-actors";
-import sources from "./sources";
 import breakpoints from "./breakpoints";
 import pendingBreakpoints from "./pending-breakpoints";
 import asyncRequests from "./async-requests";
@@ -25,7 +23,6 @@ import threads from "./threads";
 import eventListenerBreakpoints from "./event-listeners";
 
 export default {
-  sourceActors,
   breakpoints,
   pendingBreakpoints,
   asyncRequests,
