+++ src/devtools/client/debugger/src/utils/source.ts
@@ -7,7 +7,6 @@
 import { getUnicodeUrl } from "devtools/client/shared/unicode-url";
 
 import type { SymbolDeclarations } from "../reducers/ast";
-import type { SourceActor } from "../reducers/source-actors";
 import { SourceContent, MiniSource } from "ui/reducers/sources";
 import { LoadingStatus } from "ui/utils/LoadingStatus";
 
@@ -361,10 +360,6 @@ export function getMode(
   }
 
   return { name: "text" };
-}
-
-export function isInlineScript(source: SourceActor) {
-  return source.introductionType === "scriptElement";
 }
 
 type SourceContentObject = {
